% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{K nearest neighbors algo}
\usage{
knn(X.mat, y.vec, testX.vec, max.neighbors)
}
\arguments{
\item{X.mat}{numeric train feature matrix [n x p]}

\item{y.vec}{numeric train label vector [n], either all 0/1 for binary classification, or other real numbers for regression (multi-class classification not supposrted).}

\item{testX.vec}{numeric test feature vector [p]}

\item{max.neighbors}{scalar integer, max number of neighbors.}
}
\value{
numeric vector of size max neighbors, predictions from 1 to max.neighbors
}
\description{
R function that wraps our C++ code.
}
\examples{
data(zip.train, package="ElemStatLearn")
i01 <- chich(zip.train[,1] \%in\% c(0,1))
train.i <- i01[1:5]
test.i <- i01[6]
x <- zip.train[train.i, -1]
y <- zip.train[train.i, 1]
testX <- zip.train[test.i, -1]
knn(X, y, testX, 3)
zip.train[test.i, 1]
}
